Split up main file for readability
